remove rags
