+++ src/EventStore.ClusterNode/EventStore.ClusterNode.csproj
@@ -17,7 +17,6 @@
 				<ProjectReference Include="..\EventStore.Common\EventStore.Common.csproj" />
 				<ProjectReference Include="..\EventStore.Core\EventStore.Core.csproj" />
 				<ProjectReference Include="..\EventStore.Projections.Core\EventStore.Projections.Core.csproj" />
-				<ProjectReference Include="..\EventStore.Rags\EventStore.Rags.csproj" />
 		</ItemGroup>
 		<ItemGroup>
 				<None Include="..\EventStore.Common\Log\logconfig.json">
