+++ src/EventStore.Common/Configuration/OptionsDumper.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text;
-using EventStore.Rags;
 using Microsoft.Extensions.Configuration;
 
 namespace EventStore.Common.Configuration {
@@ -67,5 +66,22 @@ namespace EventStore.Common.Configuration {
 			string FormatValue(string key, string value) =>
 				string.IsNullOrEmpty(value) ? "<empty>" : sensitiveOptions.Contains(key) ? "****" : value;
 		}
+
+		internal static class NameTranslators {
+			public static string PrefixEnvironmentVariable(string name, string prefix) {
+				return prefix + CombineByPascalCase(name, "_");
+			}
+
+			public static string CombineByPascalCase(string name, string token) {
+				var regex = new System.Text.RegularExpressions.Regex(
+					@"(?<=[A-Z])(?=[A-Z][a-z])|(?<=[^A-Z])(?=[A-Z])|(?<=[A-Za-z])(?=[^A-Za-z])");
+				return regex.Replace(name, token);
+			}
+
+			public static string None(string name) {
+				return name;
+			}
+		}
+
 	}
 }
