+++ src/EventStore.Common/EventStore.Common.csproj
@@ -31,10 +31,10 @@
 		<PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
 		<PackageReference Include="Serilog.Sinks.File" Version="4.1.0" />
 		<PackageReference Include="Serilog.Extensions.Logging" Version="3.0.1" />
+		<PackageReference Include="YamlDotnet" Version="5.2.1" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\EventStore.Common.Utils\EventStore.Common.Utils.csproj" />
-		<ProjectReference Include="..\EventStore.Rags\EventStore.Rags.csproj" />
 	</ItemGroup>
 
 </Project>
