--- src/EventStore.Core.Tests/Common/EventStoreOptionsTests/TestArgs.cs
@@ -1,21 +1,0 @@
-﻿using EventStore.Common.Options;
-using System.Net;
-
-namespace EventStore.Core.Tests.Common {
-	public class TestArgs : IOptions {
-		public bool Help { get; set; }
-		public bool Version { get; set; }
-		public string Config { get; set; }
-		public string Log { get; set; }
-		public IPEndPoint[] GossipSeed { get; set; }
-		public bool WhatIf { get; set; }
-		public ProjectionType RunProjections { get; set; }
-
-		public int HttpPort { get; set; }
-
-		public TestArgs() {
-			HttpPort = 2112;
-			Log = "~/logs";
-		}
-	}
-}
