--- src/EventStore.Rags.Tests/EventStore.Rags.Tests.csproj
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net5.0</TargetFramework>
-		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-		<Platforms>x64</Platforms>
-	</PropertyGroup>
-	<ItemGroup>
-		<ProjectReference Include="..\EventStore.Rags\EventStore.Rags.csproj" />
-	</ItemGroup>
-	<ItemGroup>
-		<Content Include="YamlTests\*.yaml">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</Content>
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
-		<PackageReference Include="NUnit" Version="3.12.0" />
-		<PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
-	</ItemGroup>
-</Project>
