--- src/EventStore.Rags.Tests/YamlTests/config_with_section_and_dictionary.yaml
@@ -1,5 +1,0 @@
-﻿Name: foo
-Section:
-  Roles: 
-      'accounting': $admins
-      'it-experts': $experts
