--- src/EventStore.Rags.Tests/YamlTests/valid_config.yaml
@@ -1,1 +1,0 @@
-﻿Name: foo
