--- src/EventStore.Rags/ArgMask.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace EventStore.Rags {
-	/// <summary>
-	/// Provides the ability to mask an option when dumping the options
-	/// </summary>
-	[AttributeUsage(AttributeTargets.Property)]
-	public class ArgMask : Attribute {
-		
-	}
-}
