--- src/EventStore.Rags/EventStore.Rags.csproj
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net5.0</TargetFramework>
-		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-		<Platforms>x64</Platforms>
-	</PropertyGroup>
-	<ItemGroup>
-		<PackageReference Include="YamlDotNet" Version="5.2.1" />
-	</ItemGroup>
-</Project>
